move GroupStackRow to archive
