+++ app/src/main/java/io/taptap/stupidenglish/MainActivity.kt
@@ -586,7 +586,7 @@ fun NavController.navigateToTab(
 //7) https://stackoverflow.com/questions/67252538/jetpack-compose-update-composable-when-list-changes
 
 
-//переложи GroupStack в Archive
+
 //вынеси повторяющуюся ебалу про Groups из Import и Add
 //переверстай AddWordScreen, чтобы не прыгало ничего
 //выделение групп квадратное
